node warehouse drawer refactoring
