+++ utk_curio/frontend/urban-workflows/src/components/packs/PackCard.module.css
@@ -1,0 +1,152 @@
+/* ── Pack card ─────────────────────────────────────────────────────────── */
+
+.card {
+    display: grid;
+    grid-template-columns: 72px 1fr auto;
+    gap: 12px;
+    align-items: center;
+    padding: 12px;
+    border: 1px solid #e5e5e5;
+    border-radius: 6px;
+    background: #ffffff;
+}
+
+/* Icon tile */
+
+.cardIcon {
+    width: 72px;
+    height: 72px;
+    border-radius: 4px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 1.375rem;
+    font-weight: 700;
+    flex-shrink: 0;
+}
+
+.cardIconWarm {
+    background: #fff8ee;
+    border: 1px solid #fbaa69;
+    color: #fbaa69;
+}
+
+.cardIconCool {
+    background: #eff6ff;
+    border: 1px solid #a6c8ff;
+    color: #3e6ab2;
+}
+
+.cardIconViolet {
+    background: #f0f4ff;
+    border: 1px solid #3e6ab2;
+    color: #3e6ab2;
+}
+
+/* Body */
+
+.cardBody {
+    min-width: 0;
+}
+
+.cardTitle {
+    margin: 0 0 4px;
+    font-size: 0.875rem;
+    font-weight: 700;
+    overflow-wrap: anywhere;
+}
+
+.cardMeta {
+    margin: 0 0 8px;
+    font-size: 0.75rem;
+    color: #666666;
+}
+
+.tagRow {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 6px;
+}
+
+.tag {
+    display: inline-block;
+    padding: 2px 8px;
+    border-radius: 3px;
+    background: #f0f0f0;
+    font-size: 0.625rem;
+    color: #666666;
+}
+
+.tagUpdate {
+    background: #fff8ee;
+    border: 1px solid #fbaa69;
+    color: #7a4a12;
+    font-weight: 700;
+}
+
+/* Action column */
+
+.cardAction {
+    flex-shrink: 0;
+}
+
+.btnInstall {
+    min-width: 88px;
+    height: 32px;
+    padding: 0 12px;
+    border: none;
+    border-radius: 4px;
+    background: #1e1f23;
+    color: #ffffff;
+    font: inherit;
+    font-size: 0.8125rem;
+    font-weight: 700;
+    cursor: pointer;
+}
+
+.btnInstall:hover:not(:disabled) {
+    background: #2c2d32;
+}
+
+.btnInstallAccent {
+    background: #fbaa69;
+    color: #1e1f23;
+}
+
+.btnInstallAccent:hover:not(:disabled) {
+    background: #fb9a4f;
+}
+
+.btnInstall:disabled,
+.btnInstallAccent:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+}
+
+.btnInstalled {
+    min-width: 104px;
+    height: 32px;
+    padding: 0 12px 0 28px;
+    border: 1px solid #1e1f23;
+    border-radius: 4px;
+    background: #ffffff;
+    color: #1e1f23;
+    font: inherit;
+    font-size: 0.8125rem;
+    font-weight: 700;
+    cursor: default;
+    position: relative;
+}
+
+.btnInstalled::before {
+    content: "";
+    position: absolute;
+    left: 12px;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 6px;
+    height: 6px;
+    border-radius: 50%;
+    background: #2ba677;
+}
+
