+++ utk_curio/frontend/urban-workflows/src/components/packs/PackSearchRow.module.css
@@ -1,0 +1,57 @@
+/* ── Search + sort row ─────────────────────────────────────────────────── */
+
+.searchRow {
+    flex-shrink: 0;
+    display: flex;
+    gap: 10px;
+    padding: 0 16px 12px;
+}
+
+.searchWrap {
+    flex: 1;
+    position: relative;
+}
+
+.searchIcon {
+    position: absolute;
+    left: 12px;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 14px;
+    height: 14px;
+    color: #999999;
+    pointer-events: none;
+}
+
+.searchInput {
+    width: 100%;
+    box-sizing: border-box;
+    height: 40px;
+    padding: 0 12px 0 36px;
+    border: 1px solid #e5e5e5;
+    border-radius: 4px;
+    background: #fafafa;
+    font: inherit;
+    font-size: 0.8125rem;
+}
+
+.searchInput::placeholder {
+    color: #999999;
+}
+
+.sortSelect {
+    flex-shrink: 0;
+    width: 100px;
+    height: 40px;
+    padding: 0 28px 0 12px;
+    border: 1px solid #cccccc;
+    border-radius: 4px;
+    background: #ffffff;
+    font: inherit;
+    font-size: 0.75rem;
+    appearance: none;
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='8' height='5' viewBox='0 0 8 5'%3E%3Cpath fill='%231E1F23' d='M0 0l4 5 4-5z'/%3E%3C/svg%3E");
+    background-repeat: no-repeat;
+    background-position: right 10px center;
+}
+
