+++ utk_curio/frontend/urban-workflows/src/components/packs/packTypes.ts
@@ -1,0 +1,6 @@
+/** Tab identifiers for the Node Warehouse Drawer. */
+export type DrawerTab = "featured" | "browse" | "installed" | "updates";
+
+/** Sort modes available for the pack list. */
+export type SortMode = "new" | "name";
+
